Changes to address review comments.
