+++ metacat-common/src/main/java/com/netflix/metacat/common/json/MetacatJsonParseException.java
@@ -1,0 +1,38 @@
+/*
+ *       Copyright 2017 Netflix, Inc.
+ *          Licensed under the Apache License, Version 2.0 (the "License");
+ *          you may not use this file except in compliance with the License.
+ *          You may obtain a copy of the License at
+ *              http://www.apache.org/licenses/LICENSE-2.0
+ *          Unless required by applicable law or agreed to in writing, software
+ *          distributed under the License is distributed on an "AS IS" BASIS,
+ *          WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *          See the License for the specific language governing permissions and
+ *          limitations under the License.
+ */
+
+package com.netflix.metacat.common.json;
+
+import lombok.Getter;
+
+/**
+ * Metacat Json Parse Exception.
+ *
+ * @author zhenl
+ * @since 1.1.0
+ */
+@Getter
+public class MetacatJsonParseException extends MetacatJsonException {
+    private final String inputJson;
+
+    /**
+     * Constructor.
+     *
+     * @param inputJson input string
+     * @param message   exception message
+     */
+    public MetacatJsonParseException(final String inputJson, final String message) {
+        super(message);
+        this.inputJson = inputJson;
+    }
+}
