+++ metacat-connector-hive/src/main/java/com/netflix/metacat/connector/hive/HiveConnectorPlugin.java
@@ -20,7 +20,7 @@ import com.netflix.metacat.common.server.connectors.ConnectorFactory;
 import com.netflix.metacat.common.server.connectors.ConnectorInfoConverter;
 import com.netflix.metacat.common.server.connectors.ConnectorPlugin;
 import com.netflix.metacat.common.server.connectors.ConnectorTypeConverter;
-import com.netflix.metacat.common.server.util.MetacatConnectorConfig;
+import com.netflix.metacat.common.server.util.ConnectorConfig;
 import com.netflix.metacat.connector.hive.converters.HiveConnectorInfoConverter;
 import com.netflix.metacat.connector.hive.converters.HiveTypeConverter;
 
@@ -52,10 +52,10 @@ public class HiveConnectorPlugin implements ConnectorPlugin {
     @Override
     public ConnectorFactory create(
         final String catalogName,
-        final MetacatConnectorConfig metacatConnectorConfig
+        final ConnectorConfig connectorConfig
     ) {
         return new HiveConnectorFactory(
-            catalogName, (HiveConnectorInfoConverter) INFO_CONVERTER_HIVE, metacatConnectorConfig
+            catalogName, (HiveConnectorInfoConverter) INFO_CONVERTER_HIVE, connectorConfig
         );
     }
