+++ metacat-connector-hive/src/main/java/com/netflix/metacat/connector/hive/configs/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Created by zhenl on 6/16/17.
+ */
+package com.netflix.metacat.connector.hive.configs;
