+++ metacat-connector-s3/src/main/java/com/netflix/metacat/connector/s3/S3ConnectorPlugin.java
@@ -17,7 +17,7 @@ import com.netflix.metacat.common.server.connectors.ConnectorFactory;
 import com.netflix.metacat.common.server.connectors.ConnectorInfoConverter;
 import com.netflix.metacat.common.server.connectors.ConnectorPlugin;
 import com.netflix.metacat.common.server.connectors.ConnectorTypeConverter;
-import com.netflix.metacat.common.server.util.MetacatConnectorConfig;
+import com.netflix.metacat.common.server.util.ConnectorConfig;
 import com.netflix.metacat.common.type.TypeRegistry;
 import com.netflix.metacat.connector.pig.converters.PigTypeConverter;
 import lombok.NonNull;
@@ -50,9 +50,9 @@ public class S3ConnectorPlugin implements ConnectorPlugin {
     @Override
     public ConnectorFactory create(
         @Nonnull final String connectorName,
-        @Nonnull @NonNull final MetacatConnectorConfig metacatConnectorConfig) {
+        @Nonnull @NonNull final ConnectorConfig connectorConfig) {
         return new S3ConnectorFactory(connectorName,
-            metacatConnectorConfig.getConfiguration(), (S3ConnectorInfoConverter) getInfoConverter());
+            connectorConfig.getConfiguration(), (S3ConnectorInfoConverter) getInfoConverter());
     }
 
     /**
